corrected compiling under gcc 4.3.3
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@1924 406a0c4d-033a-0410-8de8-e80135713968

+++ src/tests.cpp
@@ -840,7 +840,7 @@ SphThreadFunc_t RwlockReader ( void * pArg )
 SphThreadFunc_t RwlockWriter ( void * pArg )
 {
 	assert ( g_tRwlock.WriteLock() );
-	g_iRwlock += int(pArg);
+	g_iRwlock += size_t(pArg);
 	sphSleepMsec ( 3 );
 	assert ( g_tRwlock.Unlock() );
 	return 0;
